community_usecase/requirements-analysis-assistant/tests/requirements.txt
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-timeout==2.2.0
pytest-xdist==3.5.0
coverage==7.4.1
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1
websockets==12.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
cryptography==41.0.7
pyyaml==6.0.1
typer==0.9.0
rich==13.7.0
fastapi==0.109.0
uvicorn==0.27.0
starlette==0.35.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==1.10.0
aiohttp-retry==2.8.3
tenacity==8.2.3
ratelimit==2.2.1
backoff==2.2.1
prometheus-client==0.19.0
structlog==24.1.0
python-json-logger==2.0.7
colorama==0.4.6
tqdm==4.66.1 